Revert "Gp2 auth0 (#1754)"
This reverts commit 3b083ea5d707c6a687e05fcf28f619fa45857e0b.
