+++ .pnp.cjs
@@ -75,10 +75,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         "reference": "workspace:apps/e2e-tests"
       },
       {
-        "name": "@asap-hub/gp2-auth-frontend",
-        "reference": "workspace:apps/gp2-auth-frontend"
-      },
-      {
         "name": "@asap-hub/gp2-frontend",
         "reference": "workspace:apps/gp2-frontend"
       },
@@ -197,8 +193,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       ["@asap-hub/fixtures", ["workspace:packages/fixtures"]],
       ["@asap-hub/flags", ["workspace:packages/flags"]],
       ["@asap-hub/frontend-utils", ["virtual:9d902e8fa3d0aec40e001519f3af8204bbaacbb7348be367280bb310f9537e3019252456410d2b71e7b1075fd4a7eb26825f777fb5c7d4d6219fc027ca539fa7#workspace:packages/frontend-utils", "workspace:packages/frontend-utils"]],
-      ["@asap-hub/gp2-auth-frontend", ["workspace:apps/gp2-auth-frontend"]],
-      ["@asap-hub/gp2-components", ["virtual:c4ae2f625a4ac5aa649abbf60db7c71aca414aadbf350d87ea37595af32595ce70907b878a98750e798e6867f5b52cf0a985bd4cd784c0ad5a3f626d3cbad95a#workspace:packages/gp2-components", "virtual:cf2d1d1368b2a34e41fdfdbccc7e758f24c8ace0b6ea4b15f22eba0a7c07dde8d57c3d9100bc914cf890c201e5db942ceb8e085e5972893e60e8732fef3e4632#workspace:packages/gp2-components", "workspace:packages/gp2-components"]],
+      ["@asap-hub/gp2-components", ["virtual:c4ae2f625a4ac5aa649abbf60db7c71aca414aadbf350d87ea37595af32595ce70907b878a98750e798e6867f5b52cf0a985bd4cd784c0ad5a3f626d3cbad95a#workspace:packages/gp2-components", "workspace:packages/gp2-components"]],
       ["@asap-hub/gp2-frontend", ["workspace:apps/gp2-frontend"]],
       ["@asap-hub/gp2-server", ["workspace:apps/gp2-server"]],
       ["@asap-hub/message-templates", ["workspace:packages/message-templates"]],
@@ -363,10 +358,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       [
         "@asap-hub/frontend-utils",
         "virtual:9d902e8fa3d0aec40e001519f3af8204bbaacbb7348be367280bb310f9537e3019252456410d2b71e7b1075fd4a7eb26825f777fb5c7d4d6219fc027ca539fa7#workspace:packages/frontend-utils"
-      ],
-      [
-        "@asap-hub/gp2-auth-frontend",
-        "workspace:apps/gp2-auth-frontend"
       ],
       [
         "@asap-hub/gp2-components",
@@ -6105,7 +6096,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       ],
       [
         "jsdom",
-        "virtual:61a64643837afd5405cc03238cfc085017b2368a7dc8660b30e3544523c3f52527959ac7ec64eb44bcb071ccdae5690543b0a4ef2dd3608212fed3cd93acc0b8#npm:16.5.3"
+        "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:16.5.3"
       ],
       [
         "jsesc",
@@ -10280,13 +10271,11 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@babel/runtime-corejs3", "npm:7.18.6"],
             ["@types/auth0-js", "npm:9.14.4"],
             ["@types/jest", "npm:26.0.24"],
-            ["@types/jsdom", "npm:16.2.3"],
             ["@types/node", "npm:14.14.45"],
             ["@types/pify", "npm:3.0.2"],
             ["auth0-js", "npm:9.16.0"],
             ["camelcase", "npm:6.2.0"],
             ["core-js", "npm:3.12.1"],
-            ["jsdom", "virtual:61a64643837afd5405cc03238cfc085017b2368a7dc8660b30e3544523c3f52527959ac7ec64eb44bcb071ccdae5690543b0a4ef2dd3608212fed3cd93acc0b8#npm:16.5.3"],
             ["pify", "npm:5.0.0"]
           ],
           "linkType": "SOFT",
@@ -10337,7 +10326,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["cross-env", "npm:7.0.3"],
             ["history", "npm:4.10.1"],
             ["identity-obj-proxy", "npm:3.0.0"],
-            ["jsdom", "virtual:61a64643837afd5405cc03238cfc085017b2368a7dc8660b30e3544523c3f52527959ac7ec64eb44bcb071ccdae5690543b0a4ef2dd3608212fed3cd93acc0b8#npm:16.5.3"],
+            ["jsdom", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:16.5.3"],
             ["react", "npm:17.0.2"],
             ["react-app-polyfill", "npm:1.0.6"],
             ["react-dom", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:17.0.2"],
@@ -10699,44 +10688,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "linkType": "SOFT",
         }]
       ]],
-      ["@asap-hub/gp2-auth-frontend", [
-        ["workspace:apps/gp2-auth-frontend", {
-          "packageLocation": "./apps/gp2-auth-frontend/",
-          "packageDependencies": [
-            ["@asap-hub/gp2-auth-frontend", "workspace:apps/gp2-auth-frontend"],
-            ["@asap-hub/auth", "workspace:packages/auth"],
-            ["@asap-hub/auth-frontend-utils", "workspace:packages/auth-frontend-utils"],
-            ["@asap-hub/dom-test-utils", "workspace:packages/dom-test-utils"],
-            ["@asap-hub/eslint-config-asap-hub", "workspace:packages/eslint-config-asap-hub"],
-            ["@asap-hub/gp2-components", "virtual:cf2d1d1368b2a34e41fdfdbccc7e758f24c8ace0b6ea4b15f22eba0a7c07dde8d57c3d9100bc914cf890c201e5db942ceb8e085e5972893e60e8732fef3e4632#workspace:packages/gp2-components"],
-            ["@asap-hub/react-components", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#workspace:packages/react-components"],
-            ["@babel/runtime-corejs3", "npm:7.18.6"],
-            ["@testing-library/dom", "npm:7.31.2"],
-            ["@testing-library/jest-dom", "npm:5.16.4"],
-            ["@testing-library/react", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:11.2.7"],
-            ["@testing-library/user-event", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:12.8.3"],
-            ["@types/history", "npm:4.7.11"],
-            ["@types/jest", "npm:26.0.24"],
-            ["@types/jsdom", "npm:16.2.3"],
-            ["@types/react", "npm:17.0.45"],
-            ["@types/react-dom", "npm:17.0.17"],
-            ["@types/react-router-dom", "npm:5.1.7"],
-            ["@types/testing-library__jest-dom", "npm:5.9.5"],
-            ["@types/testing-library__react", "npm:10.0.1"],
-            ["cross-env", "npm:7.0.3"],
-            ["history", "npm:4.10.1"],
-            ["identity-obj-proxy", "npm:3.0.0"],
-            ["jsdom", "virtual:61a64643837afd5405cc03238cfc085017b2368a7dc8660b30e3544523c3f52527959ac7ec64eb44bcb071ccdae5690543b0a4ef2dd3608212fed3cd93acc0b8#npm:16.5.3"],
-            ["react", "npm:17.0.2"],
-            ["react-app-polyfill", "npm:1.0.6"],
-            ["react-dom", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:17.0.2"],
-            ["react-router-dom", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:5.2.0"],
-            ["react-scripts", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:4.0.3"],
-            ["typescript", "patch:typescript@npm%3A4.4.4#~builtin<compat/typescript>::version=4.4.4&hash=ddd1e8"]
-          ],
-          "linkType": "SOFT",
-        }]
-      ]],
       ["@asap-hub/gp2-components", [
         ["virtual:c4ae2f625a4ac5aa649abbf60db7c71aca414aadbf350d87ea37595af32595ce70907b878a98750e798e6867f5b52cf0a985bd4cd784c0ad5a3f626d3cbad95a#workspace:packages/gp2-components", {
           "packageLocation": "./.yarn/__virtual__/@asap-hub-gp2-components-virtual-50aff566aa/1/packages/gp2-components/",
@@ -10789,69 +10740,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           ],
           "packagePeers": [
             "@asap-hub/react-context",
-            "@types/asap-hub__react-context",
-            "@types/emotion__react",
-            "@types/history",
-            "@types/react-dom",
-            "@types/react-router-dom",
-            "@types/react",
-            "history",
-            "react-dom",
-            "react-router-dom",
-            "react"
-          ],
-          "linkType": "SOFT",
-        }],
-        ["virtual:cf2d1d1368b2a34e41fdfdbccc7e758f24c8ace0b6ea4b15f22eba0a7c07dde8d57c3d9100bc914cf890c201e5db942ceb8e085e5972893e60e8732fef3e4632#workspace:packages/gp2-components", {
-          "packageLocation": "./.yarn/__virtual__/@asap-hub-gp2-components-virtual-9dc13beec3/1/packages/gp2-components/",
-          "packageDependencies": [
-            ["@asap-hub/gp2-components", "virtual:cf2d1d1368b2a34e41fdfdbccc7e758f24c8ace0b6ea4b15f22eba0a7c07dde8d57c3d9100bc914cf890c201e5db942ceb8e085e5972893e60e8732fef3e4632#workspace:packages/gp2-components"],
-            ["@asap-hub/eslint-config-asap-hub", "workspace:packages/eslint-config-asap-hub"],
-            ["@asap-hub/fixtures", "workspace:packages/fixtures"],
-            ["@asap-hub/react-components", "virtual:9d902e8fa3d0aec40e001519f3af8204bbaacbb7348be367280bb310f9537e3019252456410d2b71e7b1075fd4a7eb26825f777fb5c7d4d6219fc027ca539fa7#workspace:packages/react-components"],
-            ["@asap-hub/react-context", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#workspace:packages/react-context"],
-            ["@asap-hub/routing", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#workspace:packages/routing"],
-            ["@babel/core", "npm:7.18.6"],
-            ["@babel/preset-react", "virtual:61a64643837afd5405cc03238cfc085017b2368a7dc8660b30e3544523c3f52527959ac7ec64eb44bcb071ccdae5690543b0a4ef2dd3608212fed3cd93acc0b8#npm:7.18.6"],
-            ["@babel/runtime-corejs3", "npm:7.18.6"],
-            ["@emotion/babel-plugin", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#npm:11.3.0"],
-            ["@emotion/jest", "virtual:d0e958afa23d83fcbe1e8341a17fed5245116729be9039e4a7f723d5645a3b4a0db7dc084c7f30fa2114faa4c170592f819708d736b1e487b4ae7cd58824e005#npm:11.3.0"],
-            ["@emotion/react", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#npm:11.4.0"],
-            ["@testing-library/dom", "npm:7.31.2"],
-            ["@testing-library/jest-dom", "npm:5.16.4"],
-            ["@testing-library/react", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:11.2.7"],
-            ["@testing-library/react-hooks", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#npm:4.0.1"],
-            ["@testing-library/user-event", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:12.8.3"],
-            ["@types/asap-hub__assets", "workspace:@types/asap-hub__assets"],
-            ["@types/asap-hub__react-context", null],
-            ["@types/emotion__react", null],
-            ["@types/hast-util-sanitize", "workspace:@types/hast-util-sanitize"],
-            ["@types/history", "npm:4.7.11"],
-            ["@types/jest", "npm:26.0.24"],
-            ["@types/react", "npm:17.0.45"],
-            ["@types/react-dom", "npm:17.0.17"],
-            ["@types/react-router-dom", "npm:5.1.7"],
-            ["@types/react-router-hash-link", "npm:1.2.1"],
-            ["@types/react-test-renderer", "npm:17.0.2"],
-            ["@types/testing-library__jest-dom", "npm:5.9.5"],
-            ["@types/testing-library__react", "npm:10.0.1"],
-            ["@types/testing-library__react-hooks", "npm:3.4.1"],
-            ["@types/uuid", "npm:8.3.4"],
-            ["core-js", "npm:3.12.1"],
-            ["dom-to-playwright", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#npm:0.1.3"],
-            ["emotion-normalize", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#npm:11.0.1"],
-            ["expect-playwright", "npm:0.8.0"],
-            ["history", "npm:4.10.1"],
-            ["jest-playwright-jsdom", "npm:0.0.4"],
-            ["jest-playwright-preset", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#npm:1.7.2"],
-            ["playwright", "npm:1.16.1"],
-            ["playwright-core", "npm:1.16.1"],
-            ["react", "npm:17.0.2"],
-            ["react-dom", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:17.0.2"],
-            ["react-router-dom", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:5.2.0"],
-            ["react-test-renderer", "virtual:3a393e218825bde954376ca1a828a8b21ca2967b8d720dd56f28d8017fc081fa726c0b293069a94a55394c33a36ded19a9a4675c0d537b344c90f8add76eb926#npm:17.0.2"]
-          ],
-          "packagePeers": [
             "@types/asap-hub__react-context",
             "@types/emotion__react",
             "@types/history",
@@ -50376,10 +50264,10 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           ],
           "linkType": "SOFT",
         }],
-        ["virtual:61a64643837afd5405cc03238cfc085017b2368a7dc8660b30e3544523c3f52527959ac7ec64eb44bcb071ccdae5690543b0a4ef2dd3608212fed3cd93acc0b8#npm:16.5.3", {
-          "packageLocation": "./.yarn/__virtual__/jsdom-virtual-5b44661c87/0/cache/jsdom-npm-16.5.3-a7674d3b6b-19d107eeaa.zip/node_modules/jsdom/",
+        ["virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:16.5.3", {
+          "packageLocation": "./.yarn/__virtual__/jsdom-virtual-ebb151d51c/0/cache/jsdom-npm-16.5.3-a7674d3b6b-19d107eeaa.zip/node_modules/jsdom/",
           "packageDependencies": [
-            ["jsdom", "virtual:61a64643837afd5405cc03238cfc085017b2368a7dc8660b30e3544523c3f52527959ac7ec64eb44bcb071ccdae5690543b0a4ef2dd3608212fed3cd93acc0b8#npm:16.5.3"],
+            ["jsdom", "virtual:70ee702ae21962651e36bbfc38610149cb09f8829834f72efabf6ff499ef36bdf828f0ce38bc9269076bac68bed4fbc3b604042ba63016aed4c3a27d5308300f#npm:16.5.3"],
             ["@types/canvas", null],
             ["abab", "npm:2.0.5"],
             ["acorn", "npm:8.2.4"],
@@ -50396,7 +50284,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["nwsapi", "npm:2.2.0"],
             ["parse5", "npm:6.0.1"],
             ["request", "npm:2.88.2"],
-            ["request-promise-native", "virtual:5b44661c8763c5341cc20a4cf9e930fa9c5e1913ddfbc4738eb26e949440397bd50e3dbef6161a9ea96bdcb71b8deaa1bd16f3a8058b2f774bac0fb18bed56c7#npm:1.0.9"],
+            ["request-promise-native", "virtual:ebb151d51cbd3f60c5612cfab45939e7b1dafd2239f8fd19e4b6defcfcefcefc6e40a3919b61bd1acf0d3c83d6d175426f83b2030b405148cea646354b2ab964#npm:1.0.9"],
             ["saxes", "npm:5.0.1"],
             ["symbol-tree", "npm:3.2.4"],
             ["tough-cookie", "npm:4.0.0"],
@@ -58512,10 +58400,10 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           ],
           "linkType": "HARD",
         }],
-        ["virtual:acea8d22f30d4b31b65059d117ff6cf918326ef703d8a5a898be7333e153fcbdef7432e8a7f567745bd052fa7d919de69bf909f248dbf9924e3c8d25657835aa#npm:1.1.4", {
-          "packageLocation": "./.yarn/__virtual__/request-promise-core-virtual-6010d7a00b/0/cache/request-promise-core-npm-1.1.4-cb9fff6c90-c798bafd55.zip/node_modules/request-promise-core/",
+        ["virtual:699433408fc018228f49add4f399b1de9c2911354d19488aa0f4eb2ebda14f95241aebf4ddccb399c6a6ff6a8ee99b069b62a90a84d39261fc78c30638ad9a02#npm:1.1.4", {
+          "packageLocation": "./.yarn/__virtual__/request-promise-core-virtual-8ad97539d1/0/cache/request-promise-core-npm-1.1.4-cb9fff6c90-c798bafd55.zip/node_modules/request-promise-core/",
           "packageDependencies": [
-            ["request-promise-core", "virtual:acea8d22f30d4b31b65059d117ff6cf918326ef703d8a5a898be7333e153fcbdef7432e8a7f567745bd052fa7d919de69bf909f248dbf9924e3c8d25657835aa#npm:1.1.4"],
+            ["request-promise-core", "virtual:699433408fc018228f49add4f399b1de9c2911354d19488aa0f4eb2ebda14f95241aebf4ddccb399c6a6ff6a8ee99b069b62a90a84d39261fc78c30638ad9a02#npm:1.1.4"],
             ["@types/request", null],
             ["lodash", "npm:4.17.20"],
             ["request", "npm:2.88.2"]
@@ -58558,13 +58446,13 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           ],
           "linkType": "HARD",
         }],
-        ["virtual:5b44661c8763c5341cc20a4cf9e930fa9c5e1913ddfbc4738eb26e949440397bd50e3dbef6161a9ea96bdcb71b8deaa1bd16f3a8058b2f774bac0fb18bed56c7#npm:1.0.9", {
-          "packageLocation": "./.yarn/__virtual__/request-promise-native-virtual-acea8d22f3/0/cache/request-promise-native-npm-1.0.9-6ae8e592e8-3e2c694eef.zip/node_modules/request-promise-native/",
+        ["virtual:ebb151d51cbd3f60c5612cfab45939e7b1dafd2239f8fd19e4b6defcfcefcefc6e40a3919b61bd1acf0d3c83d6d175426f83b2030b405148cea646354b2ab964#npm:1.0.9", {
+          "packageLocation": "./.yarn/__virtual__/request-promise-native-virtual-699433408f/0/cache/request-promise-native-npm-1.0.9-6ae8e592e8-3e2c694eef.zip/node_modules/request-promise-native/",
           "packageDependencies": [
-            ["request-promise-native", "virtual:5b44661c8763c5341cc20a4cf9e930fa9c5e1913ddfbc4738eb26e949440397bd50e3dbef6161a9ea96bdcb71b8deaa1bd16f3a8058b2f774bac0fb18bed56c7#npm:1.0.9"],
+            ["request-promise-native", "virtual:ebb151d51cbd3f60c5612cfab45939e7b1dafd2239f8fd19e4b6defcfcefcefc6e40a3919b61bd1acf0d3c83d6d175426f83b2030b405148cea646354b2ab964#npm:1.0.9"],
             ["@types/request", null],
             ["request", "npm:2.88.2"],
-            ["request-promise-core", "virtual:acea8d22f30d4b31b65059d117ff6cf918326ef703d8a5a898be7333e153fcbdef7432e8a7f567745bd052fa7d919de69bf909f248dbf9924e3c8d25657835aa#npm:1.1.4"],
+            ["request-promise-core", "virtual:699433408fc018228f49add4f399b1de9c2911354d19488aa0f4eb2ebda14f95241aebf4ddccb399c6a6ff6a8ee99b069b62a90a84d39261fc78c30638ad9a02#npm:1.1.4"],
             ["stealthy-require", "npm:1.1.1"],
             ["tough-cookie", "npm:2.5.0"]
           ],
