+++ apps/crn-auth-frontend/src/utils.ts
@@ -1,0 +1,9 @@
+export const getHubUrlFromRedirect = (): string => {
+  const redirectUri = new URLSearchParams(window.location.search).get(
+    'redirect_uri',
+  );
+  if (!redirectUri) {
+    throw new Error('Redirect uri must be provided');
+  }
+  return new URL(redirectUri).origin;
+};
