--- apps/gp2-auth-frontend/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: ['@asap-hub/eslint-config-asap-hub/react'],
-  rules: {
-    'no-console': 'off',
-  },
-};
