--- apps/gp2-auth-frontend/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@asap-hub/gp2-auth-frontend",
-  "private": true,
-  "version": "0.1.0",
-  "engines": {
-    "node": "16.x"
-  },
-  "scripts": {
-    "start": "cross-env PORT=3002 TSC_COMPILE_ON_ERROR=true react-scripts start",
-    "build": "node scripts/build.js",
-    "typecheck": "tsc",
-    "test": "echo Run the workspace root script instead",
-    "eject": "react-scripts eject"
-  },
-  "dependencies": {
-    "@asap-hub/auth": "workspace:*",
-    "@asap-hub/auth-frontend-utils": "workspace:*",
-    "@asap-hub/gp2-components": "workspace:*",
-    "@asap-hub/react-components": "workspace:*",
-    "@babel/runtime-corejs3": "7.18.6",
-    "history": "^4",
-    "react": "^17.0.2",
-    "react-app-polyfill": "^1.0.6",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0"
-  },
-  "devDependencies": {
-    "@asap-hub/dom-test-utils": "workspace:*",
-    "@asap-hub/eslint-config-asap-hub": "workspace:*",
-    "@testing-library/dom": "7.31.2",
-    "@testing-library/jest-dom": "5.16.4",
-    "@testing-library/react": "11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/history": "4.7.11",
-    "@types/jest": "26.0.24",
-    "@types/jsdom": "^16",
-    "@types/react": "17.0.45",
-    "@types/react-dom": "17.0.17",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/testing-library__jest-dom": "^5.9.5",
-    "@types/testing-library__react": "^10.0.1",
-    "cross-env": "^7.0.3",
-    "identity-obj-proxy": "^3.0.0",
-    "jsdom": "^16.5.3",
-    "react-scripts": "^4.0.3",
-    "typescript": "4.4.4"
-  }
-}
