--- apps/gp2-auth-frontend/patches/react-scripts.patch
@@ -1,23 +1,0 @@
-diff --git a/config/webpack.config.js b/config/webpack.config.js
-index 25840d9..1874339 100644
---- a/config/webpack.config.js
-+++ b/config/webpack.config.js
-@@ -24,7 +24,7 @@ const ManifestPlugin = require('webpack-manifest-plugin');
- const InterpolateHtmlPlugin = require('react-dev-utils/InterpolateHtmlPlugin');
- const WorkboxWebpackPlugin = require('workbox-webpack-plugin');
- const WatchMissingNodeModulesPlugin = require('react-dev-utils/WatchMissingNodeModulesPlugin');
--const ModuleScopePlugin = require('react-dev-utils/ModuleScopePlugin');
-+// const ModuleScopePlugin = require('react-dev-utils/ModuleScopePlugin');
- const getCSSModuleLocalIdent = require('react-dev-utils/getCSSModuleLocalIdent');
- const paths = require('./paths');
- const modules = require('./modules');
-@@ -318,7 +318,8 @@ module.exports = function(webpackEnv) {
-         // To fix this, we prevent you from importing files out of src/ -- if you'd like to,
-         // please link the files into your node_modules/ and let module-resolution kick in.
-         // Make sure your source files are compiled, as they will not be processed in any way.
--        new ModuleScopePlugin(paths.appSrc, [paths.appPackageJson]),
-+        // allow files in another part of the monorepo
-+        // new ModuleScopePlugin(paths.appSrc, [paths.appPackageJson]),
-       ],
-     },
-     resolveLoader: {
