--- apps/gp2-auth-frontend/public/manifest.json
@@ -1,20 +1,0 @@
-{
-  "short_name": "GP2 Hub Signin",
-  "name": "GP2 Hub Signin",
-  "icons": [
-    {
-      "src": "favicon.png",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/png"
-    },
-    {
-      "src": "logo192.png",
-      "type": "image/png",
-      "sizes": "192x192"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
