--- apps/gp2-auth-frontend/scripts/build.js
@@ -1,9 +1,0 @@
-const { resolve } = require('path');
-const { env } = require('process');
-const { build } = require('@asap-hub/auth-frontend-utils');
-
-build(resolve(__dirname, '../build'), {
-  ...env,
-  PUBLIC_URL:
-    env.AUTH_GP2_FRONTEND_BASE_URL || 'https://dev.gp2.asap.science/.auth/',
-});
