--- apps/gp2-auth-frontend/src/auth0/__mocks__/web-auth.ts
@@ -1,3 +1,0 @@
-export const authorizeWithSso = jest.fn();
-export const authorizeWithEmailPassword = jest.fn(async () => undefined);
-export const sendPasswordResetLink = jest.fn(async () => undefined);
