--- apps/gp2-auth-frontend/src/config.ts
@@ -1,2 +1,0 @@
-export const AUTH0_CLIENT_ID = process.env.REACT_APP_GP2_AUTH0_CLIENT_ID || '';
-export const AUTH0_DOMAIN = process.env.REACT_APP_GP2_AUTH0_DOMAIN || '';
