--- apps/gp2-auth-frontend/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "composite": false,
-    "incremental": false,
-    "emitDeclarationOnly": false,
-    "noEmit": true,
-    "lib": ["esnext", "dom", "dom.iterable"],
-    "jsx": "react-jsx"
-  },
-  "include": ["src"]
-}
