+++ packages/algolia/package.json
@@ -15,8 +15,8 @@
   "dependencies": {
     "@algolia/client-common": "*",
     "@algolia/client-search": "*",
-    "@babel/runtime-corejs3": "7.18.6",
     "algoliasearch": "^4.13.1",
+    "@babel/runtime-corejs3": "7.18.6",
     "prettier": "^2.7.1"
   },
   "devDependencies": {
