+++ packages/auth-frontend-utils/package.json
@@ -17,7 +17,6 @@
     "auth0-js": "^9.16.0",
     "camelcase": "^6.2.0",
     "core-js": "^3.12.1",
-    "jsdom": "^16.5.3",
     "pify": "^5.0.0"
   },
   "devDependencies": {
@@ -28,7 +27,6 @@
     "@babel/preset-react": "7.18.6",
     "@types/auth0-js": "^9.14.4",
     "@types/jest": "26.0.24",
-    "@types/jsdom": "^16",
     "@types/node": "^14.14.45",
     "@types/pify": "^3.0.2"
   }
