+++ packages/auth-frontend-utils/src/index.ts
@@ -1,13 +1,2 @@
 export * from './errors';
 export * from './web-auth';
-export * from './build';
-
-export const getHubUrlFromRedirect = (): string => {
-  const redirectUri = new URLSearchParams(window.location.search).get(
-    'redirect_uri',
-  );
-  if (!redirectUri) {
-    throw new Error('Redirect uri must be provided');
-  }
-  return new URL(redirectUri).origin;
-};
