+++ packages/gp2-components/src/index.ts
@@ -1,7 +1,2 @@
 export { NavigationHeader } from './organism';
-export {
-  Layout,
-  DashboardPage,
-  DashboardPageBody,
-  BasicLayout,
-} from './templates';
+export { Layout, DashboardPage, DashboardPageBody } from './templates';
