+++ packages/gp2-components/src/templates/index.ts
@@ -1,4 +1,3 @@
 export { default as Layout } from './Layout';
 export { default as DashboardPage } from './DashboardPage';
 export { default as DashboardPageBody } from './DashboardPageBody';
-export { default as BasicLayout } from './BasicLayout';
